Extract drag detection into class
